+++ src/frontends/qt4/GuiWorkArea_Private.h
@@ -126,15 +126,18 @@ struct GuiWorkArea::Private
 		delete screen_;
 		pixel_ratio_ = p->pixelRatio();
 		if (lyxrc.use_qimage) {
-			QImage *x = new QImage(pixel_ratio_ * p->viewport()->width(),
-				pixel_ratio_ * p->viewport()->height(), QImage::Format_ARGB32_Premultiplied);
+			QImage *x = 
+				new QImage(static_cast<int>(pixel_ratio_ * p->viewport()->width()),
+						   static_cast<int>(pixel_ratio_ * p->viewport()->height()),
+						   QImage::Format_ARGB32_Premultiplied);
 #if QT_VERSION >= 0x050000
 			x->setDevicePixelRatio(pixel_ratio_);
 #endif
 			screen_ = x;
 		} else {
-			QPixmap *x = new QPixmap(pixel_ratio_ * p->viewport()->width(),
-				pixel_ratio_ * p->viewport()->height());
+			QPixmap *x = 
+				new QPixmap(static_cast<int>(pixel_ratio_ * p->viewport()->width()),
+							static_cast<int>(pixel_ratio_ * p->viewport()->height()));
 #if QT_VERSION >= 0x050000
 			x->setDevicePixelRatio(pixel_ratio_);
 #endif
